+++ src/main/java/com/ebook/ebook/entity/Book.java
@@ -2,19 +2,21 @@ package com.ebook.ebook.entity;
 
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 
+import java.math.BigDecimal;
+
 public class Book {
     private Long book_id;
     private String isbn;
     private String name;
     private String author;
     private String press;
-    private Double price;
+    private BigDecimal price;
     private Integer inventory;
     private String description;
     private String image;
-    public Book(Long book_id,String isbn,String name,
-                String author,String press,Double price,
-                Integer inventory,String description,String image)
+    public Book(Long book_id, String isbn, String name,
+                String author, String press, BigDecimal price,
+                Integer inventory, String description, String image)
     {
         this.book_id=book_id;
         this.isbn=isbn;
@@ -28,7 +30,7 @@ public class Book {
     }
 
     public Book(Long book_id,String isbn,String name,
-                String author,String press,Double price)
+                String author,String press,BigDecimal price)
     {
         this.book_id=book_id;
         this.isbn=isbn;
@@ -43,7 +45,7 @@ public class Book {
     public String getName() {return this.name;}
     public String getAuthor() {return this.author;}
     public String getPress() {return this.press;}
-    public Double getPrice() {return this.price;}
+    public BigDecimal getPrice() {return this.price;}
     public Integer getInventory() {return this.inventory;}
     public String getDescription() {return this.description;}
     public String getImage() {return this.image;}
